Split supervisor and machine code into separate binaries
