+++ src/pmptest.rs
@@ -1,8 +1,9 @@
 
+use rvirt::*;
 use crate::constants::SYMBOL_PA2VA_OFFSET;
 use crate::fdt::*;
 use crate::pagedebug;
-use crate::pmap;
+use crate::pmap::{self, BootPageTable};
 use crate::pmp;
 use crate::print;
 use crate::trap::constants::*;
@@ -21,7 +22,7 @@ pub unsafe fn pmptest_mstart(hartid: u64, device_tree_blob: u64) {
     csrs!(medeleg, 0xb1ff);
     csrw!(mie, 0x888);
     csrs!(mstatus, STATUS_MPP_S);
-    csrw!(mepc, sstart as u64);
+    // csrw!(mepc, sstart as u64); -- TODO!!!!!!!!!!!!!!1
     csrw!(mcounteren, 0xffffffff);
     csrw!(mscratch, 0x80800000 + 0x1000 * (hartid+1));
 
@@ -30,10 +31,11 @@ pub unsafe fn pmptest_mstart(hartid: u64, device_tree_blob: u64) {
          ::: "t0"  : "volatile");
 
     // Minimal page table to boot into S mode.
-    *((pmap::mboot_page_table_pa()) as *mut u64) = 0x00000000 | 0xcf;
-    *((pmap::mboot_page_table_pa()+16) as *mut u64) = 0x20000000 | 0xcf;
-    *((pmap::mboot_page_table_pa()+4088) as *mut u64) = 0x20000000 | 0xcf;
-    csrw!(satp, 8 << 60 | (pmap::mboot_page_table_pa() >> 12));
+    let boot_page_table_pa = SHARED_STATICS.boot_page_table.as_ptr() as u64;
+    *((boot_page_table_pa) as *mut u64) = 0x00000000 | 0xcf;
+    *((boot_page_table_pa+16) as *mut u64) = 0x20000000 | 0xcf;
+    *((boot_page_table_pa+4088) as *mut u64) = 0x20000000 | 0xcf;
+    csrw!(satp, 8 << 60 | (boot_page_table_pa >> 12));
 
     // Text segment
     pmp::install_pmp_napot(0, pmp::LOCK | pmp::READ | pmp::EXEC, 0x80000000, 0x200000);
@@ -55,29 +57,29 @@ pub unsafe fn pmptest_mstart(hartid: u64, device_tree_blob: u64) {
           mret" :: "r"(device_tree_blob) : "a0", "a1" : "volatile");
 }
 
-unsafe fn sstart(device_tree_blob: u64) {
-    asm!("li t0, $0
-          add sp, sp, t0" :: "i"(SYMBOL_PA2VA_OFFSET)  : "t0" : "volatile");
-    csrw!(stvec, (||{panic!("Trap on hart 0?!")}) as fn() as *const () as u64);
+// unsafe fn sstart(device_tree_blob: u64) {
+//     asm!("li t0, $0
+//           add sp, sp, t0" :: "i"(SYMBOL_PA2VA_OFFSET)  : "t0" : "volatile");
+//     csrw!(stvec, (||{panic!("Trap on hart 0?!")}) as fn() as *const () as u64);
 
-    // Read and process host FDT.
-    let fdt = Fdt::new(device_tree_blob);
-    assert!(fdt.magic_valid());
-    assert!(fdt.version() >= 17 && fdt.last_comp_version() <= 17);
-    assert!(fdt.total_size() < 64 * 1024);
-    let machine = fdt.parse();
+//     // Read and process host FDT.
+//     let fdt = Fdt::new(device_tree_blob);
+//     assert!(fdt.magic_valid());
+//     assert!(fdt.version() >= 17 && fdt.last_comp_version() <= 17);
+//     assert!(fdt.total_size() < 64 * 1024);
+//     let machine = fdt.parse();
 
-    // Initialize UART
-    if let Some(ty) = machine.uart_type {
-        print::UART_WRITER.lock().init(machine.uart_address, ty);
-    }
+//     // Initialize UART
+//     if let Some(ty) = machine.uart_type {
+//         print::UART_WRITER.lock().init(machine.uart_address, ty);
+//     }
 
-    println!("ONLINE");
+//     println!("ONLINE");
 
-    loop {}
-}
+//     loop {}
+// }
 
-#[allow(unused)]
-unsafe fn ustart() {
+// #[allow(unused)]
+// unsafe fn ustart() {
 
-}
+// }
